Move gamma stuff to separate files
