+++ tools/run_tests/xds_k8s_test_driver/framework/infrastructure/traffic_director.py
@@ -831,10 +831,6 @@ class TrafficDirectorAppNetManager(TrafficDirectorManager):
         super().cleanup(force=force)
 
 
-class TrafficDirectorGammaManager(TrafficDirectorManager):
-    """Gamma."""
-
-
 class TrafficDirectorSecureManager(TrafficDirectorManager):
     SERVER_TLS_POLICY_NAME = "server-tls-policy"
     CLIENT_TLS_POLICY_NAME = "client-tls-policy"
